+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Technical - Make the CI lint the commit messages.
 - Readme - Add the test coverage badge.
 - Branch command - Add `--help` option.
+- Branch command - Add branch creation.
+- Technical - Add an api error deserializer.
 
 ### Changed
 - Technical - Untrack IDE `.idea` folder.